Allow to set ES heap memory
fixes #15
helps with #5

We set Heap to system memory/2 with a maximum of 30 (given by Java) and
a minimum of 1 (arbitraty but sensible). You can always override the
setting because it's just a default. But we enforce gigabyte as unit.

Maybe the calcultaion could be written easier. I'm always open for
suggestions. I hoped the current variant would be at least easier to
understand than more "elegant" ones.